apps/ai-service/requirements.txt
# Core dependencies
python-dotenv>=1.1.0,<2.0.0
black>=25.1.0,<26.0.0
isort>=6.0.1,<7.0.0

# LangChain and LangGraph
langchain>=0.3.22,<0.4.0
langchain-core>=0.3.76,<0.4.0
langgraph>=0.3.21,<0.4.0
grandalf>=0.8,<0.9
langchain-ollama>=0.3.0,<0.4.0
langchain-openai>=0.3.11,<0.4.0
langchain-experimental>=0.3.4,<0.4.0
langchain-google-genai>=2.1.4,<3.0.0
langgraph-cli[inmem]>=0.2.10,<0.3.0
# langgraph-supervisor>=0.0.21

# Web frameworks
streamlit>=1.44.1,<2.0.0
fastapi>=0.115.12,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0

# Vector DB and search
faiss-cpu>=1.10.0,<2.0.0
rank-bm25>=0.2.2,<0.3.0

# ML and transformers
numpy>=1.29
torch>=1.13.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.3.0,<3.0.0
tqdm>=4.67.1,<5.0.0
openpyxl>=3.1.5,<4.0.0

# Database
motor>=3.7.0,<4.0.0
asyncpg>=0.30.0,<0.31.0
pymongo>=4.13.2,<5.0.0

# API and validation
pydantic>=2.11.3,<3.0.0
pydantic-settings>=2.8.1,<3.0.0
python-multipart>=0.0.20,<0.0.21
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
email-validator>=2.3.0,<3.0.0

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.11.7,<0.12.0
pytesseract>=0.3.13,<0.4.0
pillow>=11.3.0,<12.0.0
pdf2image>=1.17.0,<2.0.0
opencv-python>=4.12.0.88,<5.0.0.0
docling>=2.23.1,<3.0.0

# NLP
nltk>=3.8.1

# Visualization
plotly>=6.3.0,<7.0.0

# Graph algorithms
python-louvain>=0.16,<0.17
networkx>=3.0,<4.0

# Utilities
requests>=2.32.0,<3.0.0
typer>=0.15.4,<0.16.0
pylint>=3.3.7,<4.0.0
# mcp>=1.9.0,<2.0.0
